# src/main/java/com/andreyk/practiceproject/api/controller/AuthController.java
package com.andreyk.practiceproject.api.controller;

import com.andreyk.practiceproject.api.dto.auth.LoginRequest;
import com.andreyk.practiceproject.api.dto.auth.LoginResponse;
import com.andreyk.practiceproject.api.dto.auth.RefreshRequest;
import com.andreyk.practiceproject.core.service.AuthService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/auth")
@RequiredArgsConstructor
public class AuthController {

    private final AuthService authService;

    @PostMapping("/login")
    public ResponseEntity<LoginResponse> authenticate(@RequestBody LoginRequest request) {
        return ResponseEntity.ok(authService.authenticate(request));
    }

    @PostMapping("/refresh")
    public ResponseEntity<LoginResponse> refreshToken(@RequestBody RefreshRequest request) {
        return ResponseEntity.ok(authService.refreshToken(request));
    }
}