# src/main/java/com/andreyk/practiceproject/api/dto/auth/LoginRequest.java
package com.andreyk.practiceproject.api.dto.auth;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class LoginRequest {
    @JsonProperty("username")
    private String username;

    @JsonProperty("password")
    private String password;
}